+++ service/src/main/java/com/itellyou/service/user/star/impl/UserStarServiceImpl.java
@@ -12,6 +12,7 @@ import com.itellyou.service.common.StarService;
 import com.itellyou.service.event.OperationalPublisher;
 import com.itellyou.service.user.UserInfoService;
 import com.itellyou.service.user.UserSingleService;
+import com.itellyou.service.user.star.UserStarSingleService;
 import com.itellyou.util.DateUtils;
 import com.itellyou.util.RedisUtils;
 import org.slf4j.Logger;
@@ -37,14 +38,16 @@ public class UserStarServiceImpl implements StarService<UserStarModel> {
     private final UserStarDao starDao;
     private final UserInfoService infoService;
     private final UserSingleService singleService;
+    private final UserStarSingleService starSingleService;
 
     private final OperationalPublisher operationalPublisher;
 
     @Autowired
-    public UserStarServiceImpl(UserStarDao starDao, UserInfoService infoService, UserSingleService singleService, OperationalPublisher operationalPublisher){
+    public UserStarServiceImpl(UserStarDao starDao, UserInfoService infoService, UserSingleService singleService, UserStarSingleService starSingleService, OperationalPublisher operationalPublisher){
         this.starDao = starDao;
         this.infoService = infoService;
         this.singleService = singleService;
+        this.starSingleService = starSingleService;
         this.operationalPublisher = operationalPublisher;
     }
 
@@ -53,6 +56,8 @@ public class UserStarServiceImpl implements StarService<UserStarModel> {
     @Caching(evict = {  @CacheEvict(key = "T(String).valueOf(#model.userId).concat('-').concat(#model.createdUserId).concat('-detail')") , @CacheEvict(key = "T(String).valueOf(#model.userId).concat('-').concat(#model.createdUserId)") })
     public int insert(UserStarModel model) throws Exception {
         try{
+            UserStarModel starModel = starSingleService.find(model.getUserId(),model.getCreatedUserId());
+            if(starModel != null) return 1;
             int result = starDao.insert(model);
             if(result != 1) throw new Exception("写入关注记录失败");
             result = infoService.updateStarCount(model.getCreatedUserId(),1);
@@ -76,6 +81,8 @@ public class UserStarServiceImpl implements StarService<UserStarModel> {
         UserInfoModel infoModel = singleService.findById(userId);
         try{
             if(infoModel == null) throw new Exception("错误的用户ID");
+            UserStarModel starModel = starSingleService.find(userId,followerId);
+            if(starModel == null) return 1;
             int result = starDao.delete(userId,followerId);
             if(result != 1) throw new Exception("删除关注记录失败");
             result = infoService.updateStarCount(followerId,-1);
